Remove eslint consistent return

+++ server/controllers/tasksController.js
@@ -12,7 +12,6 @@ module.exports = {
             const column = board.columns.id(columnId);
 
             if (column) {
-                console.log(column);
                 const {tasks} = column;
                 return res.json(tasks);
             }
@@ -29,7 +28,6 @@ module.exports = {
             // console.log(req.params);
             const {taskName} = req.body;
             const priority = 'low'; // default for a newly created task
-            console.log(taskName);
             const board = await Board.findById(boardId);
 
             if (!board) {
@@ -42,13 +40,12 @@ module.exports = {
                 return res.status(404).json({error: 'Column not found'});
             }
 
-            column.tasks.push({taskName: taskName, priority: priority});
+            column.tasks.push({taskName, priority});
 
             board.save();
 
             const task = column.tasks[column.tasks.length - 1];
 
-            console.log(task);
             return res.json(task);
         } catch (error) {
             console.error(error);
@@ -61,8 +58,6 @@ module.exports = {
         try {
             const {boardId, columnId, taskId} = req.params;
             const {taskName, priority, taskDetail, assignedUserIds} = req.body;
-            // const board = await Board.findById(boardId)
-            console.log(boardId);
 
             const board = await Board.findById(boardId);
 
@@ -103,10 +98,27 @@ module.exports = {
             const {boardId, columnId, taskId} = req.params;
 
             const board = await Board.findById(boardId);
+
+            if (!board) {
+                return res.status(404).json({error: 'Board not found'});
+            }
+
             const column = board.columns.id(columnId);
+
+            if (!column) {
+                return res.status(404).json({error: 'Column not found'});
+            }
+
+            const task = column.tasks.id(taskId);
+
+            if (!task) {
+                return res.status(404).json({error: 'Task not found'});
+            }
+
             column.tasks.id(taskId).deleteOne();
+
             await board.save();
-            console.log(board);
+
             return res.json(board);
         } catch (error) {
             console.error(error);
